fix(getPages): support new Downloader page
For performance reasons David Revoy updated the Downloader page to
require that a langauge pack be requested before it would be zipped
into an archive file for download. As a result, the `wget` command
in the `getPages.bash` script was retrieiving an HTML document with
links to cached archive files, instead of the archive file itself.

Now prime the cache by making a GET request for the desired language
and then download that langauge archive file using `wget`.

Furthermore, enable `--continue` so that if the script is interrupted,
`wget` can continue where it left off, further reducing the load on
the remote server.

See related commit - https://framagit.org/peppercarrot/website/commit/03c38afba2138ca8fc8063b0889c50f7b9914808

Signed-off-by: Hutson Betts <hutson@hyper-expanse.net>