docs(README): formatted for clarity
From my experience getting setup with this project, I modified the
layout, and initial setup instructions, to be a little more
accessible.

Several changes are quite subjective, but hopefully will prove valuable
to others in helping to smooth the on-boarding process for people who
want to contribute to, or create their own, books based on Pepper&Carrot.